@@ -1,10 +1,8 @@
 module github.com/rsteube/carapace
 
-go 1.14
+go 1.15
 
 require (
-	github.com/mitchellh/go-ps v1.0.0
-	github.com/spf13/cobra v1.2.1
+	github.com/spf13/cobra v1.6.1
 	github.com/spf13/pflag v1.0.5
-	golang.org/x/sys v0.0.0-20210510120138-977fb7262007
 )
