@@ -1,19 +1,45 @@
 package carapace
 
 import (
+	"encoding/json"
 	"fmt"
+	"os"
 	"sort"
 	"testing"
+	"time"
 
 	"github.com/rsteube/carapace/internal/assert"
 	"github.com/rsteube/carapace/internal/common"
+	"github.com/rsteube/carapace/pkg/style"
 )
+
+func init() {
+	os.Unsetenv("LS_COLORS")
+}
 
 func assertEqual(t *testing.T, expected, actual InvokedAction) {
 	sort.Sort(common.ByValue(expected.rawValues))
 	sort.Sort(common.ByValue(actual.rawValues))
 
-	assert.Equal(t, fmt.Sprintf("%+v\n", expected), fmt.Sprintf("%+v\n", actual))
+	e, _ := json.MarshalIndent(expected.rawValues, "", "  ")
+	a, _ := json.MarshalIndent(actual.rawValues, "", "  ")
+	assert.Equal(t, string(e), string(a))
+
+	eMeta, _ := json.MarshalIndent(expected.meta, "", "  ")
+	aMeta, _ := json.MarshalIndent(actual.meta, "", "  ")
+	assert.Equal(t, string(eMeta), string(aMeta))
+}
+
+func assertNotEqual(t *testing.T, expected, actual InvokedAction) {
+	sort.Sort(common.ByValue(expected.rawValues))
+	sort.Sort(common.ByValue(actual.rawValues))
+
+	e, _ := json.MarshalIndent(expected.rawValues, "", "  ")
+	a, _ := json.MarshalIndent(actual.rawValues, "", "  ")
+
+	if string(e) == string(a) {
+		t.Errorf("should differ:\n%v", a)
+	}
 }
 
 func TestActionCallback(t *testing.T) {
@@ -25,12 +51,28 @@ func TestActionCallback(t *testing.T) {
 		})
 	})
 	expected := InvokedAction{
-		rawValues: common.RawValuesFrom("a", "b", "c"),
-		nospace:   false,
-		skipcache: false,
+		Action{
+			rawValues: common.RawValuesFrom("a", "b", "c"),
+		},
 	}
 	actual := a.Invoke(Context{})
 	assertEqual(t, expected, actual)
+}
+
+func TestCache(t *testing.T) {
+	f := func() Action {
+		return ActionCallback(func(c Context) Action {
+			return ActionValues(time.Now().String())
+		}).Cache(15 * time.Millisecond)
+	}
+
+	a1 := f().Invoke(Context{})
+	a2 := f().Invoke(Context{})
+	assertEqual(t, a1, a2)
+
+	time.Sleep(16 * time.Millisecond)
+	a3 := f().Invoke(Context{})
+	assertNotEqual(t, a1, a3)
 }
 
 func TestSkipCache(t *testing.T) {
@@ -44,11 +86,11 @@ func TestSkipCache(t *testing.T) {
 			Suffix("").
 			ToA()
 	})
-	if a.skipcache {
-		t.Fatal("uninvoked skipcache should be false")
+	if !a.meta.Messages.IsEmpty() {
+		t.Fatal("uninvoked action should not contain messages")
 	}
-	if !a.Invoke(Context{}).skipcache {
-		t.Fatal("invoked skipcache should be true")
+	if a.Invoke(Context{}).meta.Messages.IsEmpty() {
+		t.Fatal("invoked action should contain messages")
 	}
 }
 
@@ -63,58 +105,127 @@ func TestNoSpace(t *testing.T) {
 			Suffix("").
 			ToA()
 	})
-	if a.nospace {
-		t.Fatal("uninvoked nospace should be false")
+	if a.meta.Nospace.Matches("x") {
+		t.Fatal("uninvoked nospace should not match")
 	}
-	if !a.Invoke(Context{}).nospace {
-		t.Fatal("invoked nospace should be true")
+	if !a.Invoke(Context{}).meta.Nospace.Matches("x") {
+		t.Fatal("invoked nospace should match")
 	}
 }
 
 func TestActionDirectories(t *testing.T) {
 	assertEqual(t,
-		ActionValues("example/", "docs/", "internal/", "pkg/").noSpace(true).Invoke(Context{}),
-		ActionDirectories().Invoke(Context{CallbackValue: ""}),
+		ActionStyledValues(
+			"example/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"docs/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"internal/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"pkg/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"third_party/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+		).NoSpace('/').Invoke(Context{}),
+		ActionDirectories().Invoke(Context{CallbackValue: ""}).Filter([]string{"vendor/"}),
 	)
 
 	assertEqual(t,
-		ActionValues("example/", "docs/", "internal/", "pkg/").noSpace(true).Invoke(Context{}).Prefix("./"),
-		ActionDirectories().Invoke(Context{CallbackValue: "./"}),
+		ActionStyledValues(
+			"example/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"docs/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"internal/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"pkg/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"third_party/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+		).NoSpace('/').Invoke(Context{}).Prefix("./"),
+		ActionDirectories().Invoke(Context{CallbackValue: "./"}).Filter([]string{"./vendor/"}),
 	)
 
 	assertEqual(t,
-		ActionValues("_test/", "cmd/").noSpace(true).Invoke(Context{}).Prefix("example/"),
+		ActionStyledValues(
+			"_test/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"cmd/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+		).NoSpace('/').Invoke(Context{}).Prefix("example/"),
 		ActionDirectories().Invoke(Context{CallbackValue: "example/"}),
 	)
 
 	assertEqual(t,
-		ActionValues("_test/", "cmd/").noSpace(true).Invoke(Context{}).Prefix("example/"),
+		ActionStyledValues(
+			"cmd/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+		).NoSpace('/').Invoke(Context{}).Prefix("example/"),
 		ActionDirectories().Invoke(Context{CallbackValue: "example/cm"}),
 	)
 }
 
 func TestActionFiles(t *testing.T) {
 	assertEqual(t,
-		ActionValues("README.md", "example/", "docs/", "internal/", "pkg/").noSpace(true).Invoke(Context{}),
-		ActionFiles(".md").Invoke(Context{CallbackValue: ""}),
+		ActionStyledValues(
+			"README.md", style.Of("fg-default", "bg-default"),
+			"example/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"docs/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"internal/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"pkg/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"third_party/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+		).NoSpace('/').Invoke(Context{}),
+		ActionFiles(".md").Invoke(Context{CallbackValue: ""}).Filter([]string{"vendor/"}),
 	)
 
 	assertEqual(t,
-		ActionValues("_test/", "cmd/", "main.go", "main_test.go").noSpace(true).Invoke(Context{}).Prefix("example/"),
-		ActionFiles().Invoke(Context{CallbackValue: "example/"}),
+		ActionStyledValues(
+			"README.md", style.Of("fg-default", "bg-default"),
+			"_test/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"cmd/", style.Of("fg-default", "bg-default", style.Blue, style.Bold),
+			"main.go", style.Of("fg-default", "bg-default"),
+			"main_test.go", style.Of("fg-default", "bg-default"),
+		).NoSpace('/').Invoke(Context{}).Prefix("example/"),
+		ActionFiles().Invoke(Context{CallbackValue: "example/"}).Filter([]string{"example/example"}),
 	)
 }
 
-func TestActionMessage(t *testing.T) {
+func TestActionFilesChdir(t *testing.T) {
+	oldWd, _ := os.Getwd()
+
 	assertEqual(t,
-		ActionValuesDescribed("_", "", "ERR", "example message").noSpace(true).skipCache(true).Invoke(Context{}).Prefix("docs/"),
+		ActionMessage(fmt.Sprintf("stat %v: no such file or directory", wd("nonexistent"))).NoSpace('/').Invoke(Context{}),
+		ActionFiles(".md").Chdir("nonexistent").Invoke(Context{}),
+	)
+
+	assertEqual(t,
+		ActionMessage(fmt.Sprintf("not a directory: %v/go.mod", wd(""))).Invoke(Context{}),
+		ActionFiles(".md").Chdir("go.mod").Invoke(Context{CallbackValue: ""}),
+	)
+
+	assertEqual(t,
+		ActionStyledValues(
+			"action.go", style.Of("fg-default", "bg-default"),
+			"snippet.go", style.Of("fg-default", "bg-default"),
+		).NoSpace('/').Invoke(Context{}).Prefix("elvish/"),
+		ActionFiles().Chdir("internal/shell").Invoke(Context{CallbackValue: "elvish/"}),
+	)
+
+	if newWd, _ := os.Getwd(); oldWd != newWd {
+		t.Error("workdir should not be changed")
+	}
+}
+
+func TestActionMessage(t *testing.T) {
+	expected := ActionValues().NoSpace()
+	expected.meta.Messages.Add("example message")
+
+	assertEqual(t,
+		expected.Invoke(Context{}),
 		ActionMessage("example message").Invoke(Context{CallbackValue: "docs/"}),
+	)
+}
+
+func TestActionMessageSuppress(t *testing.T) {
+	assertEqual(t,
+		Batch(
+			ActionMessage("example message").Suppress("example"),
+			ActionValues("test"),
+		).ToA().Invoke(Context{}),
+		ActionValues("test").NoSpace('*').Invoke(Context{}), // TODO suppress does not reset nospace (is that even possible?)
 	)
 }
 
 func TestActionExecCommand(t *testing.T) {
 	assertEqual(t,
-		ActionMessage("go unknown: unknown command").noSpace(true).skipCache(true).Invoke(Context{}).Prefix("docs/"),
+		ActionMessage("go unknown: unknown command").NoSpace('/').Invoke(Context{}).Prefix("docs/"),
 		ActionExecCommand("go", "unknown")(func(output []byte) Action { return ActionValues() }).Invoke(Context{CallbackValue: "docs/"}),
 	)
 
