@@ -2,15 +2,17 @@ package carapace
 
 import "sync"
 
-type batch []Action
-type invokedBatch []InvokedAction
+type (
+	batch        []Action
+	invokedBatch []InvokedAction
+)
 
-// Batch creates a batch of Actions that can be invoked in parallel
+// Batch creates a batch of Actions that can be invoked in parallel.
 func Batch(actions ...Action) batch {
 	return batch(actions)
 }
 
-// Invoke invokes contained Actions of the batch using groutines
+// Invoke invokes contained Actions of the batch using goroutines.
 func (b batch) Invoke(c Context) invokedBatch {
 	invokedActions := make([]InvokedAction, len(b))
 	functions := make([]func(), len(b))
@@ -26,17 +28,18 @@ func (b batch) Invoke(c Context) invokedBatch {
 	return invokedActions
 }
 
-// ToA converts the batch to a implicitly merged action which is a shortcut for:
-//   ActionCallback(func(c Context) Action {
-//	 	return batch.Invoke(c).Merge().ToA()
-//	 })
+// ToA converts the batch to an implicitly merged action which is a shortcut for:
+//
+//	ActionCallback(func(c Context) Action {
+//		return batch.Invoke(c).Merge().ToA()
+//	})
 func (b batch) ToA() Action {
 	return ActionCallback(func(c Context) Action {
 		return b.Invoke(c).Merge().ToA()
 	})
 }
 
-// Merge merges Actions of a batch
+// Merge merges Actions of a batch.
 func (b invokedBatch) Merge() InvokedAction {
 	switch len(b) {
 	case 0:
