@@ -1,6 +1,8 @@
 package main
 
-import "github.com/rsteube/carapace/example/cmd"
+import (
+	"github.com/rsteube/carapace/example/cmd"
+)
 
 func main() {
 	_ = cmd.Execute()
